% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActivityData.R
\docType{data}
\name{ActivityData}
\alias{ActivityData}
\title{Activity Data Template}
\format{
\subsection{\code{ActivityData}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<ghgtools.io>
}
\usage{
ActivityData
}
\description{
Details...
}
\keyword{datasets}
