% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{check_load_data}
\alias{check_load_data}
\title{check_load_data}
\usage{
check_load_data()
}
\value{
The required dataframes for ghgtools
}
\description{
This function checks your input data and loads it into the global environment.
}
