% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{create_templates}
\alias{create_templates}
\title{create_templates}
\usage{
create_templates()
}
\value{
The activity data, asset portfolio templates, and the full emission factor library
}
\description{
This function takes \code{\link{ActivityData}}, \code{\link{AssetPortfolio}}, and \code{\link{EFLibrary}} from ghgtools and creates a csv file for each in your project directory.
}
