% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecat_lookup.R
\docType{data}
\name{Ecat_lookup}
\alias{Ecat_lookup}
\title{Emission Category Lookup}
\format{
\subsection{\code{Ecat_lookup}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<ghgtools.io>
}
\usage{
Ecat_lookup
}
\description{
Details...
}
\keyword{datasets}
