% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGRIDlookup.R
\docType{data}
\name{eGRIDlookup}
\alias{eGRIDlookup}
\title{eGRID Subregion Lookup}
\format{
\subsection{\code{eGRIDlookup}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<ghgtools.io>
}
\usage{
eGRIDlookup
}
\description{
Details...
}
\keyword{datasets}
