% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{export_efl}
\alias{export_efl}
\title{export_efl}
\usage{
export_efl(GWP, EFL = EFLibrary)
}
\arguments{
\item{GWP}{Select your GWPs. Enter with quotation marks in the function. Choices are "SAR", "AR4", "AR5", or "AR6"}

\item{EFL}{Emissions Factor Library organized according to the template. Defaults to \code{EFLibrary}}
}
\value{
The consolidated emission factor library
}
\description{
This function consolidates and returns an emission factor library based on your selection of global warming potentials (GWPs). If you have no preference for GWPs, we recommend following UNFCCC guidelines, which require the use of GWP values from the IPCC's Fifth Assessment Report (AR5).
}
\seealso{
\href{https://www.epa.gov/ghgemissions/understanding-global-warming-potentials}{Gudiance from EPA about GWPs}
}
