% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHG_RawData.R
\name{GHG_AnnualReport}
\alias{GHG_AnnualReport}
\title{GHG Annual Report}
\usage{
GHG_AnnualReport(ReportingYear)
}
\arguments{
\item{ReportingYear}{Select the year you wish to generate the GHG report for.}
}
\value{
This function is a component of the GHG Inventory Report repository.
}
\description{
GHG_Report requires a \strong{GHG_RawData.csv} file to exist in your repository. This file is generated using the \emph{GHG_RawData} function. Please visit \href{https://www.ghgtools.io/tools/}{ghgtools.io} for detailed instructions and resources.
}
