% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{ghg_inventory}
\alias{ghg_inventory}
\title{ghg_inventory}
\usage{
ghg_inventory()
}
\value{
The GHG raw data
}
\description{
Calculates a GHG inventory using input data from EF_Library, Asset_Portfolio, and Activity_Data variables in the global environment.
}
