% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHG_Inventory.R
\name{GHGinventory}
\alias{GHGinventory}
\title{ghgtools}
\usage{
GHGinventory(Activity_Data)
}
\arguments{
\item{Activity_Data}{Energy & Activity Data}
}
\value{
The GHG emissions of the activity
}
\description{
Open Source GHG Accounting Tools
}
\examples{
GHG1 <- GHGinventory(50);
GHG2 <- GHGinventory( c(50, 63, 23) );
}
