% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWPs.R
\docType{data}
\name{GWPs}
\alias{GWPs}
\title{Global Warming Potentials}
\format{
\subsection{\code{GWPs}}{

A data frame with 132 rows and 6 columns:
\describe{
\item{ghg}{Greenhouse gas}
\item{SAR}{IPCC second assessment report GWPs}
\item{TAR}{IPCC third assessment report GWPs}
\item{AR4}{IPCC fourth assessment report GWPs}
\item{AR5}{IPCC fifth assessment report GWPs}
\item{source}{The source of GWP data}
...
}
}
}
\usage{
GWPs
}
\description{
A table of all global warming potentials for greenhouse gasses included in the EFLibrary
}
\keyword{datasets}
