% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{initiate}
\alias{initiate}
\title{initiate}
\usage{
initiate(GWP)
}
\arguments{
\item{GWP}{Select your desired global warming potentials (GWPs). Enter with quotation marks in the function. Choices are "SAR", "AR4", "AR5", or "AR6"}
}
\value{
Need to fill
}
\description{
Need to fill
}
