% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data()
}
\value{
The required dataframes for ghgtools
}
\description{
This function loads data from .csv files in your working directory and creates dataframes in your global environment.
}
