% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssetPortfolio.R
\docType{data}
\name{AssetPortfolio}
\alias{AssetPortfolio}
\title{Asset Portfolio Template}
\format{
\subsection{\code{AssetPortfolio}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<ghgtools.io>
}
\usage{
AssetPortfolio
}
\description{
Details...
}
\keyword{datasets}
