% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{create_templates}
\alias{create_templates}
\title{create_templates}
\usage{
create_templates()
}
\value{
The activity data and asset portfolio templates
}
\description{
This function creates the activity data and asset portfolio templates as CSV files in your project directory.
}
