% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFLibrary.R
\docType{data}
\name{EFLibrary}
\alias{EFLibrary}
\title{Emission Factor Library}
\format{
\subsection{\code{EFLibrary}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<ghgtools.io>
}
\usage{
EFLibrary
}
\description{
Details...
}
\keyword{datasets}
