% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eGRIDlookup.R
\docType{data}
\name{eGRIDlookup}
\alias{eGRIDlookup}
\title{eGRID Subregion Lookup}
\format{
\subsection{\code{eGRIDlookup}}{

A data frame with 3 rows and 41536 columns:
\describe{
\item{zip}{zip code}
\item{service_type}{The service type, i.e. electricity}
\item{subregion}{eGRID subregion}
...
}
}
}
\usage{
eGRIDlookup
}
\description{
A table from eGRID which associated zip codes with subregion
}
\keyword{datasets}
