% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{ghg_inventory}
\alias{ghg_inventory}
\title{ghg_inventory}
\usage{
ghg_inventory()
}
\value{
Two data sets are written to a csv file in your repository.

The first is an error report for records of activity data that failed to calculate GHG emissions properly. Please review this file carefully and make any necessary changes to your activity data and/or asset portfolio.

The second data set contains the GHG emissions for each record of activity data that passed through the function successfully.
}
\description{
ghg_inventory requires correctly formatted \strong{Activity Data, Asset Portfolio, and Emission Factor Library} to exist in your r project directory.
}
