% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghg_inventory.R
\name{ghg_inventory}
\alias{ghg_inventory}
\title{ghg_inventory}
\usage{
ghg_inventory(AD, AP, GWP, EFL = EFLibrary)
}
\arguments{
\item{AD}{Activity Data organized according to the template.}

\item{AP}{Asset Portfolio organized according the template.}

\item{GWP}{Select your GWPs. Enter with quotation marks in the function. Choices are "SAR", "TAR", "AR4", or "AR5".}

\item{EFL}{Emissions Factor Library organized according to the template. Defaults to \code{EFLibrary}}
}
\value{
The GHG raw data
}
\description{
Calculates a GHG inventory using input data from EF_Library, Asset_Portfolio, and Activity_Data variables in the global environment.
}
\examples{
AR5_Inventory <- ghg_inventory(ActivityData, AssetPortfolio, "ar5")
head(AR5_Inventory)

}
\seealso{
\href{https://www.epa.gov/ghgemissions/understanding-global-warming-potentials}{Gudiance from EPA about GWPs}
}
