% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHG_Inventory.R
\name{GHG_RawData}
\alias{GHG_RawData}
\title{ghgtools}
\usage{
GHG_RawData(GWP)
}
\arguments{
\item{GWP}{Select your desired global warming potentials (GWPs). Enter with quotation marks in the function. Choices are "SAR", "AR4", "AR5", or "AR6"}
}
\value{
Two data sets are written to a csv file in your repository.

The first is an error report for records of activity data that failed to calculate GHG emissions properly. Please review this file carefully and make any necessary changes to your activity data and/or asset portfolio.

The second data set contains the GHG emissions for each record of activity data that passed through the function successfully.
}
\description{
GHG_RawData requires correctly formatted Activity Data, Asset Portfolio, and Emission Factor Library to exist in your repository. Please visit ghgtools.io for detailed instructions and resources.
}
