% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHG_Inventory.R
\name{GHGinventory}
\alias{GHGinventory}
\title{GHG_Inventory}
\usage{
GHGinventory(Activity_Data)
}
\arguments{
\item{Activity_Data}{Energy & Activity Data}
}
\value{
The GHG emissions of the activity
}
\description{
Calculate a GHG Inventory
}
\examples{
GHG1 <- GHG_Inventory(50);
GHG2 <- GHGInventory( c(50, 63, 23) );
}
