% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWPs.R
\docType{data}
\name{GWPs}
\alias{GWPs}
\title{Global Warming Potentials}
\format{
\subsection{\code{GWPs}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<ghgtools.io>
}
\usage{
GWPs
}
\description{
Details...
}
\keyword{datasets}
